Move polyfills into vendor
