+++ src/components/error-summary/error-summary.js
@@ -1,4 +1,4 @@
-import '../../globals/polyfills/Event' // addEventListener
+import '../../vendor/polyfills/Event' // addEventListener
 
 function ErrorSummary ($module) {
   this.$module = $module
